Don't set the tmpfs size
If we do not set a size, it defaults to 1/2 of the system memory.
However cgroups still limit the usage of tmpfs.
We are allowed to roughly use 2 the cgroup memory limit in bytes.

Ideally, we should set it to half of that, which we can do in
a subsequent patch.

Signed-off-by: Mrunal Patel <mrunalp@gmail.com>