allocate configData dynamically
The content of config files is read into a buffer of fixed size
(currently CONFIGSZ is 65536 bytes) which imposes a limit on the
size of config files that systemdhook is able to parse. The patch
removes this limitation by allocating a buffer dynamically, based
on the actual config file size. The patch also fixes parse error
'trailing garbage' which could occur due to a missing NULL byte
that is expected to mark the end of the JSON string in the buffer.

As a 'bonus', the patch also fixes incorrect initialization of the
'errbuf' and 'stateData' variables.